configure: Fix clang version check for weird versions
The clang version check assumed that the version itself will always be in
the third field of the output, but there are versions that prefix the
version itself with another string, for instance as 'Ubuntu clang version
XX'. Fix this by using a proper regex to match the actual version instead
of relying on the field number to extract it.

Fixes: a8d19a67e768 ("configure: Check clang version and bail if it's <10")
Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>